feat: add env vars for INCLUDE_ENGINES and MAX_CONCURRENT_SYNC (#13)
