+++ .gitignore
@@ -4,3 +4,4 @@
 /coverage
 .eslintcache
 .env
+/test-srts
