+++ src/generateAutosubsyncSubtitles.ts
@@ -1,21 +1,18 @@
-import { access } from 'fs/promises';
 import { basename, dirname, join } from 'path';
 import { execPromise, ProcessingResult } from './helpers';
+import { existsSync } from 'fs';
 
 export async function generateAutosubsyncSubtitles(srtPath: string, videoPath: string): Promise<ProcessingResult> {
   const directory = dirname(srtPath);
   const srtBaseName = basename(srtPath, '.srt');
   const outputPath = join(directory, `${srtBaseName}.autosubsync.srt`);
 
-  // Check if synced subtitle already exists
-  try {
-    await access(outputPath);
+  const exists = existsSync(outputPath);
+  if (exists) {
     return {
       success: true,
       message: `Skipping ${outputPath} - already processed`,
     };
-  } catch (error) {
-    // File doesn't exist, proceed with sync
   }
 
   try {
